requirements/base.txt
django==5.1.1
psycopg2-binary==2.9.9
python-environ==0.4.54
django-cors-headers==4.4.0
drf-yasg==1.21.7 
djangorestframework==3.15.2 
django-filter==24.3
django-modeltranslation==0.19.9
django-redis==5.4.0
cryptography==41.0.7
celery==5.3.6
redis==5.0.1
flower==2.0.1
pillow==10.4.0
moviepy==1.0.3
djangorestframework-simplejwt==5.3.1
